Remove magic offset from replay extension.

+++ base/src/main/java/racecontrol/client/extension/contact/ContactExtension.java
@@ -87,7 +87,7 @@ public class ContactExtension
 
     public void onAccident(BroadcastingEvent event) {
         int sessionTime = client.getModel().getSessionInfo().getSessionTime();
-        int replaytime = replayOffsetExtension.getReplayTimeFromSessionTime((int) sessionTime);
+        int replaytime = replayOffsetExtension.getReplayTimeFromSessionTime(sessionTime);
         String logMessage = "Contact: " + client.getModel().getCar(event.getCarId()).getCarNumberString()
                 + "\t\t" + TimeUtils.asDuration(sessionTime)
                 + "\t" + TimeUtils.asDuration(replaytime);
